+++ src/app/engineerings/creating/layout.tsx
@@ -6,7 +6,7 @@ export const metadata = {
   openGraph: {
     title: 'テーブル作成',
     description: 'テーブルを作成するSQL文を生成します',
-    url: 'https://k8o.me/sql-statement/creating',
+    url: 'https://k8o.me/engineerings/creating',
   },
   twitter: {
     title: 'テーブル作成',
