+++ src/app/engineerings/page.tsx
@@ -4,7 +4,7 @@ export default function Page() {
   return (
     <div className="flex flex-col gap-4">
       <AppCard
-        link="/sql-statement/creating"
+        link="/engineerings/creating"
         emotion="📄"
         title="テーブル作成"
         description="テーブルを作成するSQL文を生成します"
